Revert "Demoing zero-copy save. (#567)"
This reverts commit 4b3864c802d727be3cd67e5107a0f873d047ae69.
