+++ bindings/python/Cargo.toml
@@ -9,13 +9,8 @@ rust-version = "1.74"
 name = "safetensors_rust"
 crate-type = ["cdylib"]
 
-[features]
-py38 = ["pyo3/abi3-py38"]
-py311 = ["pyo3/abi3-py311"]
-default = ["py311"]
-
 [dependencies]
-pyo3 = { version = "0.23", features = ["abi3"] }
+pyo3 = { version = "0.23", features = ["abi3", "abi3-py38"] }
 memmap2 = "0.9"
 serde_json = "1.0"
